chore: move validate.ts to middleware
